Refactored the Tree hierarchy into a single class
